requirements.txt
requests==2.31.0
Pillow==10.2.0
customtkinter==5.2.2
python-dotenv==1.0.0
pynput==1.7.6
pyautogui==0.9.54
keyboard==0.13.5
openai==1.12.0
httpx>=0.24.1
pyinstaller==6.4.0
pystray==0.19.5 